Add features to DualPortBlockRam; update verilator, gtest, cmake, etc.
DualPortBlockRam was refactored somewhat to include more features, like resettable data outputs and byte-wide write enable.  This required an update to Verilator, which required an update to MSYS2, which updated cmake and allowed the adoption of Verilator's "verilate()" cmake function for a much better build system design.  Due to the update to Verilator, many new lint errors were fixed as well.